Removing unused code example and reindexing others
